fix: Remove the Datadog logging adapter (#229)
This change removes the Datadog logging adapter. The adapter is only
consumed by `go-datalog-lib` and should not have been implemented in
`go-logger`.

With the V2 changes to `dd-trace-go` removing the sooner than later will
make migration to OpenTelemetry or `dd-trace-go` V2 simpler.

The functionality is copied into `go-datadog-lib`, see
[coopnorge/go-datadog-lib#571][coopnorge/go-datadog-lib#571] for more
information.

[coopnorge/go-datadog-lib#571]: https://github.com/coopnorge/go-datadog-lib/pull/571